refactor fstr, shaps
